refactoring: renamed TicketPurchaseForm to CTicketPurchaseForm
