+++ frontend/src/components/Customer/CTicketPurchaseView/CTicketPurchaseForm.vue
